+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/dcql/DCQLSdJwtCredentialMetadataAndValidityConstraints.kt
@@ -23,7 +23,7 @@ data class DCQLSdJwtCredentialMetadataAndValidityConstraints(
     }
 
     fun validate(actualCredentialType: String?): KmmResult<Unit> = catching {
-        if(vctValues != null && actualCredentialType !in vctValues) {
+        if (vctValues != null && actualCredentialType !in vctValues) {
             throw IllegalArgumentException("Incompatible MDOC document type.")
         }
     }